+++ frontend/src/components/user/User.ts
@@ -1,0 +1,4 @@
+export default interface User {
+  id: string;
+  username: string;
+}
