+++ frontend/src/components/user/UserPaner.tsx
@@ -1,0 +1,9 @@
+import { Typography } from "@mui/material";
+
+export default function UserPaner() {
+  return (
+    <Typography variant="h6" color="inherit" noWrap>
+      UserPaner
+    </Typography>
+  );
+}
