+++ frontend/src/components/user/index.ts
@@ -1,0 +1,1 @@
+export type { default as User } from "./User";
