+++ frontend/src/lib/socket.ts
@@ -1,12 +1,4 @@
 import { io, Socket } from "socket.io-client";
-import store, {
-  getUsers,
-  userConnect,
-  UserOnline,
-  Users,
-  Message,
-  addMessage,
-} from "./store";
 import { ServerToClientEvents, ClientToServerEvents } from "./events";
 const socket: Socket<ServerToClientEvents, ClientToServerEvents> = io(
   "http://localhost:8000",
@@ -16,9 +8,8 @@ const socket: Socket<ServerToClientEvents, ClientToServerEvents> = io(
   }
 );
 
-const { dispatch } = store;
-
 socket.on("connect_error", (err) => {
+  console.log("Socket connection error");
   console.log(err);
 });
