+++ frontend/src/lib/store.ts
@@ -1,0 +1,10 @@
+import { configureStore, createSlice, PayloadAction } from "@reduxjs/toolkit";
+
+const store = configureStore({
+  reducer: {},
+});
+
+export default store;
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
