+++ frontend/src/pages/index.tsx
@@ -1,8 +1,10 @@
 import type { NextPage } from "next";
+import { useRouter } from "next/router";
 import { useState, useEffect } from "react";
+import Head from "next/head";
+import Image from "next/image";
+import styles from "../styles/Home.module.css";
 import socket from "@/lib/socket";
-import OnlinePanel from "@/components/online_panel";
-import ChatMessages from "@/components/chat_messages";
 
 const User: NextPage = () => {
   const [message, setMessage] = useState("");
@@ -12,12 +14,12 @@ const User: NextPage = () => {
       socket.disconnect();
     };
   }, []);
+  const onClink = () => {};
   return (
     <div>
       <>
-        <ChatMessages />
         <input value={message} onChange={(e) => setMessage(e.target.value)} />
-        <OnlinePanel />
+        <button onClick={onClink}>Send</button>
       </>
     </div>
   );
